STY: Put interpolation codes to a new cpp file
